@@ -1,8 +1,56 @@
+[project]
+name = "rispy"
+license = {file = "LICENSE"}
+authors = [
+    {name = "Maik Derstappen", email = "md@derico.de"}
+]
+maintainers = [
+    {name = "Andy Shapiro", email = "shapiromatron@gmail.com"}
+]
+readme = "README.rst"
+dynamic = ["version", "description"]
+keywords = ["RIS", "parser", "bibliograph"]
+classifiers = [
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+]
+requires-python = ">=3.8"
+
+[project.urls]
+Source = "https://github.com/mrtango/rispy"
+
+[project.optional-dependencies]
+dev = [
+    "pytest ~=7.4.0",
+    "flit ~= 3.9.0",
+    "black ~= 23.7.0",
+    "ruff ~= 0.0.278",
+    "coverage ~= 7.2.7",
+]
+
 [build-system]
-requires = ["setuptools>=43", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["flit_core >=3.2,<4"]
+build-backend = "flit_core.buildapi"
 
 [tool.black]
 line-length = 98
-target-version = ['py36', 'py37', 'py38', 'py39']
+target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
 include = '\.pyi?$'
+
+[tool.ruff]
+line-length = 100
+target-version = "py38"
+select = ["F", "E", "W", "I", "UP", "S", "B", "T20", "RUF"]
+
+[tool.ruff.per-file-ignores]
+"test_*.py" = ["S101"]
+
+[tool.coverage.run]
+omit = [
+    "tests/*",
+]
