@@ -1,12 +1,11 @@
 """RIS Writer."""
 
 import warnings
-from typing import Dict, List, TextIO, Optional
 from abc import ABC
+from typing import ClassVar, Dict, List, Optional, TextIO, Type
 
 from .config import LIST_TYPE_TAGS, TAG_KEY_MAPPING
 from .utils import invert_dictionary
-
 
 __all__ = ["dump", "dumps", "BaseWriter", "RisWriter"]
 
@@ -41,7 +40,7 @@ class BaseWriter(ABC):
     START_TAG: str
     END_TAG: str = "ER"
     PATTERN: str
-    DEFAULT_IGNORE: List[str] = []
+    DEFAULT_IGNORE: ClassVar[List[str]] = []
     DEFAULT_MAPPING: Dict
     DEFAULT_LIST_TAGS: List[str]
     DEFAULT_REFERENCE_TYPE: str = "JOUR"
@@ -99,7 +98,7 @@ class BaseWriter(ABC):
             lines.append(header)
         lines.append(self._format_line(self.START_TAG, self._get_reference_type(ref)))
 
-        tags_to_skip = [self.START_TAG] + self.ignore
+        tags_to_skip = [self.START_TAG, *self.ignore]
         if self.skip_unknown_tags:
             tags_to_skip.append("UK")
 
@@ -108,7 +107,7 @@ class BaseWriter(ABC):
             try:
                 tag = self._rev_mapping[label.lower()]
             except KeyError:
-                warnings.warn(UserWarning(f"label `{label}` not exported"))
+                warnings.warn(UserWarning(f"label `{label}` not exported"), stacklevel=2)
                 continue
 
             # ignore
@@ -162,7 +161,7 @@ class RisWriter(BaseWriter):
     DEFAULT_LIST_TAGS = LIST_TYPE_TAGS
 
     def set_header(self, count):
-        return "{i}.".format(i=count)
+        return f"{count}."
 
 
 def dump(
@@ -190,7 +189,9 @@ def dump(
     file.writelines(text)
 
 
-def dumps(references: List[Dict], *, implementation: Optional[BaseWriter] = None, **kw) -> str:
+def dumps(
+    references: List[Dict], *, implementation: Optional[Type[BaseWriter]] = None, **kw
+) -> str:
     """Return an RIS formatted string.
 
     Entries are codified as dictionaries whose keys are the
